+++ src/app/app.component.scss
@@ -1,4 +1,4 @@
-@import './styles/variables';
+@import 'variables';
 
 .wrapper {
   width: 100%;
