+++ src/app/app.module.ts
@@ -15,20 +15,20 @@ import { MatTabsModule } from '@angular/material/tabs';
 import { MatMenuModule } from '@angular/material/menu';
 import { MatFormFieldModule } from '@angular/material/form-field';
 import { HeaderComponent } from './header/header.component';
-import { ProfileSidebarComponent } from './profile-sidebar/profile-sidebar.component';
-import { WishCardComponent } from './wish-card/wish-card.component';
-import { SearchInputComponent } from './search-input/search-input.component';
-import { ModalWindowComponent } from './modal-window/modal-window.component';
-import { EmptyFieldComponent } from './empty-field/empty-field.component';
+import { ProfileSidebarComponent } from './main/profile/profile-sidebar/profile-sidebar.component';
+import { WishCardComponent } from './main/main-page/main-page-content/wish-card/wish-card.component';
+import { SearchInputComponent } from './main/main-menu/search-input/search-input.component';
+import { ModalWindowComponent } from './main/main-page/main-page-content/wish-card/modal-window/modal-window.component';
+import { EmptyFieldComponent } from './main/main-page/main-page-content/empty-field/empty-field.component';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-import { WishcardModalComponent } from './wishcard-modal/wishcard-modal.component';
-import { ProfileEditFormComponent } from './profile-edit-form/profile-edit-form.component';
+import { WishcardModalComponent } from './main/main-page/main-page-content/wish-card/wishcard-modal/wishcard-modal.component';
+import { ProfileEditFormComponent } from './main/profile/profile-edit-form/profile-edit-form.component';
 import { LoginFormComponent } from './login-form/login-form.component';
 import { RegisterFormComponent } from './register-form/register-form.component';
 import { RouterModule, Routes } from '@angular/router';
-import { MainPageContentComponent } from './main-page-content/main-page-content.component';
-import { MainPageComponent } from './main-page/main-page.component';
-import { AddWishBlockComponent } from './add-wish-block/add-wish-block.component';
+import { MainPageContentComponent } from './main/main-page/main-page-content/main-page-content.component';
+import { MainPageComponent } from './main/main-page/main-page.component';
+import { AddWishBlockComponent } from './main/main-menu/add-wish-block/add-wish-block.component';
 import { ErrorPageComponent } from './error-page/error-page.component';
 
 const appRoutes: Routes = [
