+++ src/app/header/header.component.scss
@@ -1,4 +1,4 @@
-@import '../styles/variables';
+@import 'variables';
 
 .header {
   min-width: 100%;
