+++ src/app/login-form/login-form.component.scss
@@ -1,4 +1,4 @@
-@import '../styles/variables';
+@import 'variables';
 
 :host {
   display: flex;
