+++ src/app/main/main-menu/search-input/search-input.component.scss
@@ -1,4 +1,4 @@
-@import '../styles/variables';
+@import 'variables';
 
 .global-search-input {
   background-color: $primaryBcgColor;
