+++ src/app/main/main-menu/search-input/search-input.component.spec.ts
