+++ src/app/main/main-page/main-page-content/empty-field/empty-field.component.scss
