+++ src/app/main/main-page/main-page-content/main-page-content.component.scss
@@ -1,4 +1,4 @@
-@import '../styles/variables';
+@import 'variables';
 
 :host {
   min-width: 75%;
