+++ src/app/main/main-page/main-page-content/wish-card/modal-window/modal-window.component.ts
