+++ src/app/main/main-page/main-page-content/wish-card/wish-card.component.scss
@@ -1,4 +1,4 @@
-@import '../styles/variables';
+@import 'variables';
 
 :host {
   max-width: 30%;
