+++ src/app/main/main-page/main-page-content/wish-card/wish-card.component.spec.ts
