+++ src/app/main/main-page/main-page-content/wish-card/wishcard-modal/wishcard-modal.component.scss
@@ -1,4 +1,4 @@
-@import '../styles/variables';
+@import 'variables';
 
 .mat-dialog-title {
   display: flex;
