+++ src/app/main/main-page/main-page-content/wish-card/wishcard-modal/wishcard-modal.component.ts
