+++ src/app/main/main-page/main-page.component.scss
