+++ src/app/main/profile/profile-sidebar/profile-sidebar.component.scss
@@ -1,4 +1,4 @@
-@import '../styles/variables';
+@import 'variables';
 
 :host {
   width: 25%;
@@ -13,7 +13,7 @@
   display: flex;
   justify-self: flex-start;
   align-items: center;
-  background: url("../../../assets/images/bcgImg2.png") 100% 100% no-repeat, url("../../../assets/images/bcgImg.png") 100% 100% no-repeat;
+  background: url("../../../../assets/images/bcgImg2.png") 100% 100% no-repeat, url("../../../../assets/images/bcgImg.png") 100% 100% no-repeat;
   background-size: cover, cover;
 }
