+++ src/app/register-form/register-form.component.scss
@@ -1,4 +1,4 @@
-@import '../styles/variables';
+@import 'variables';
 
 :host {
   display: flex;
